Add files via upload
The "Block" script is just to give the Unity standard cube a little more interest(color) while I was developing the code. It is not needed for the code to work. 

The "Gravity Gun" script uses a raycast to find its target. The raycast is looking for a gameobject with a tag of "Block". You can change this in the "DoRay()" method.